lab4

+++ routes/web.php
@@ -3,7 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\postsController;
 use App\Http\Controllers\desController;
-use App\Http\Controllers\AuthorController;
+use App\Http\Controllers\usersController;
 
 
 /*
@@ -25,18 +25,22 @@ Route::get('/viewPage',[postsController::class , "viewPosts"])->name("viewPage")
 
 Route::get('viewPage/{id}', [postsController::class , "viewdes"])->name('despage');
 
-Route::get('create', [postsController::class , "create"])->name('create');
+Route::get('create', [postsController::class , "create"])->name('create')->middleware('auth');
 
-Route::get('update/{id}', [postsController::class , "update"])->name('update');
+Route::get('update/{id}', [postsController::class , "update"])->name('update')->middleware('auth');
 
-Route::put('edit/{id}', [postsController::class , "edit"])->name('edit');
+Route::put('edit/{id}', [postsController::class , "edit"])->name('edit')->middleware('auth');
 
-Route::post('store', [postsController::class , "store"])->name('store');
+Route::post('store', [postsController::class , "store"])->name('store')->middleware('auth');
 
-Route::delete('/viewPage/delete/{id}', [postsController::class , "destroy"])->name('destroy');
+Route::delete('/viewPage/delete/{id}', [postsController::class , "destroy"])->name('destroy')->middleware('auth');
 
 
-Route::get('/viewAuthor',[AuthorController::class , "viewAuthor"])->name("viewAuthor");
+Route::get('/viewusers',[usersController::class , "viewusers"])->name("viewusers");
 
-Route::get('viewAuthor/{id}', [AuthorController::class , "viewAuthorDetails"])->name('viewAuthorDetails');
+Route::get('viewusers/{id}', [usersController::class , "viewusersDetails"])->name('viewusersDetails');
 
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
